1-1347: extract feature naming pattern info to own component
